pyproject.toml
[project]
name = "nq-bot"
version = "1.1.0"
description = "nq-bot"
readme = "README.md"
requires-python = ">=3.8, <4.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_BitTorrent", "nonebot-plugin-person"]
plugin_dirs = []
builtin_plugins = ["echo"]

[tool.poetry]
name = "nq-bot"
version = "1.1.0"
description = "一个自娱自乐的产物"
authors = ["ilksese <leeyrur@gmail.com>"]
readme = "README.md"
#packages = [{include = "nq-bot"}]

[tool.poetry.dependencies]
python = "^3.10"
nonebot-adapter-onebot = "^2.2.3"
nonebot2 = {extras = ["fastapi"], version = "^2.0.1"}
aiohttp = "^3.8.5"
nonebot-plugin-bittorrent = "^0.0.13"
playwright = "^1.36.0"
nonebot-plugin-person = "^0.0.5"
pygtrans = "^1.5.3"
markdown = "^3.4.4"
requests = "^2.31.0"
requests-toolbelt = "^1.0.0"
requests-cache = "^1.1.0"
pillow = "^10.0.0"
brotli = "^1.0.9"
httpx = "^0.24.1"
objprint = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
